pyproject.toml
[project]
name = "polars-prover"
version = "0.1.0"
description = "Polars based provers for zkVMs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "loguru>=0.7.3",
  "numpy>=2.2.4",
  "polars>=1.27.0",
  "python-dotenv>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cairo/src/prover"]

[dependency-groups]
dev = ["jupyter>=1.1.1"]

[tool.isort]
profile = "black"
